moved test data & test aardvark instance into merlin.support package.  This allows client software to use the data and modules for testing.
