+++ merlin/support/data/specs/nirs_q_tags_nir_AND_sr.json
