+++ merlin/support/data/specs/swir1s_q_tags_swir1_AND_sr.json
