+++ merlin/support/data/specs/swir2s_q_tags_swir2_AND_sr.json
