+++ merlin/support/data/specs/thermals_q_tags_bt_AND_thermal_AND_NOT_tirs2.json
