+++ test/test_dates.py
@@ -5,10 +5,6 @@ def test_to_ordinal():
     assert d.to_ordinal('1999-01-01') == 729755
     assert d.to_ordinal('1999/11/01') == 730059
 
-    # This value floats.  Increments by one every day so it must not be
-    # interpreting it as a fixed day.
-    # assert d.to_ordinal('2001') == 730641
-
 
 def test_startdate():
     assert d.startdate('1980-01-01/1982-01-01') == '1980-01-01'
